% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.aggregate_ptlg_data}
\alias{.aggregate_ptlg_data}
\title{Aggregiert Daten basierend auf der gewählten Variante}
\usage{
.aggregate_ptlg_data(data_list, target_tables = 2)
}
\description{
Aggregiert Daten basierend auf der gewählten Variante
}
\keyword{internal}
