% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.export_ptlg_to_csv}
\alias{.export_ptlg_to_csv}
\title{Speichert eine Liste von Dataframes als einzelne CSV-Dateien}
\usage{
.export_ptlg_to_csv(data_list, path)
}
\description{
Speichert eine Liste von Dataframes als einzelne CSV-Dateien
}
\keyword{internal}
