% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.harmonize_ptlg_data}
\alias{.harmonize_ptlg_data}
\title{Interne Funktion zur Harmonisierung der Spaltennamen}
\usage{
.harmonize_ptlg_data(data)
}
\description{
Interne Funktion zur Harmonisierung der Spaltennamen
}
\keyword{internal}
