% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.load_ptlg_data_raw}
\alias{.load_ptlg_data_raw}
\title{Lädt Rohdaten aus einem angegebenen Pfad}
\usage{
.load_ptlg_data_raw(path)
}
\arguments{
\item{path}{Pfad zum Ordner, in dem die CSV-Dateien liegen.}
}
\value{
Eine Liste mit den eingelesenen Dataframes.
}
\description{
Lädt Rohdaten aus einem angegebenen Pfad
}
\keyword{internal}
