% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.process_variant_1}
\alias{.process_variant_1}
\title{Verarbeitet Variante 1: Master-Tabelle}
\usage{
.process_variant_1(h_data)
}
\description{
Verarbeitet Variante 1: Master-Tabelle
}
\keyword{internal}
