% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.process_variant_2}
\alias{.process_variant_2}
\title{Variante 2: Zwei Tabellen als Output}
\usage{
.process_variant_2(h_data)
}
\description{
Variante 2: Zwei Tabellen als Output
}
\keyword{internal}
