% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.process_variant_3}
\alias{.process_variant_3}
\title{Variante 3: Drei Tabellen als Outputs}
\usage{
.process_variant_3(h_data)
}
\description{
Variante 3: Drei Tabellen als Outputs
}
\keyword{internal}
