% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{.process_variant_4}
\alias{.process_variant_4}
\title{Variante 4: Vier Tabellen als Output}
\usage{
.process_variant_4(h_data)
}
\description{
Variante 4: Vier Tabellen als Output
}
\keyword{internal}
