% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlg_config.R
\docType{data}
\name{.ptlg_config}
\alias{.ptlg_config}
\title{Interne Konfiguration für PTLG-Datensätze}
\format{
An object of class \code{list} of length 2.
}
\usage{
.ptlg_config
}
\description{
Interne Konfiguration für PTLG-Datensätze
}
\keyword{internal}
