% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ptlg.R
\name{transform_data}
\alias{transform_data}
\title{Konsolidiert PTLG-Daten (4 verschiedene Optionen)}
\usage{
transform_data(path_input, target_tables = 3, path_output = path_input)
}
\arguments{
\item{path_input}{Pfad, wo die ursprünglichen Tabellen liegen.}

\item{target_tables}{Anzahl Tabellen, die ausgegeben werden sollen (1, 2, 3, oder 4). Standard: 3.}

\item{path_output}{Pfad für den Export der CSV-Dateien. Standard: gleich path_input.}
}
\value{
Schreibt die aggregierten Tabellen als CSV-Dateien auf die Platte.
}
\description{
Konsolidiert PTLG-Daten (4 verschiedene Optionen)
}
